Rewrite lxd installation action
The setup-lxd action was failing on arm64 runner,
at the iptables changes step.
This is because the arm64 runner did not have docker installed,
but the action command failed if docker was not installed.

As the setup-lxd action is listed as alpha state,
and the action is very small,
it seems fine to inline the commands here
so we can manage it ourselves to suit our use case better.

- ensure lxd latest/stable is installed
- switch the daemon group to adm (runner user on both machines is in this group)
- update iptables rules to work with docker (ignore errors if docker is not installed)

Differences between runner machines we need to handle:

`Ubuntu_ARM64_4C_16G_01` runner:
- lxd snap 5.0/stable is pre installed
- runner user is in the lxd group
- docker is not installed

`ubuntu-latest` runner:
- lxd is not installed
- runner user is not in the lxd group
- docker is installed